+++ src/Command/UpdateDataCommand.php
@@ -4,6 +4,8 @@ namespace Aislandener\MixTelematicsLaravel\Command;
 
 use Aislandener\MixTelematicsLaravel\Facades\MixTelematics;
 use Aislandener\MixTelematicsLaravel\Models\Driver;
+use Carbon\Carbon;
+use Carbon\CarbonImmutable;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Log;
 
@@ -13,14 +15,15 @@ class UpdateDataCommand extends Command
     protected $signature = "mix-telematics:populate 
                                     {--d|driver : Populate to Driver} 
                                     {--a|asset : Populate to Asset} 
-                                    {--g|group : Populate to Group} 
+                                    {--g|group : Populate to Group}
+                                    {--p|position : Populate to Position} 
                                     {--A|all : Populate all tables}";
 
     protected $description = 'Collect data in MiX Telematics and save in database';
 
     public function handle()
     {
-        if(!(collect($this->options())->only(['all', 'driver', 'asset', 'group'])->some(true))) {
+        if(!(collect($this->options())->only(['all', 'driver', 'asset', 'group', 'position'])->some(true))) {
             $this->error("Need some parameter");
             return;
         }
@@ -36,6 +39,10 @@ class UpdateDataCommand extends Command
 
         if($this->option('group') || $this->option('all')) {
             $this->populateGroup();
+        }
+
+        if($this->option('position') || $this->option('all')) {
+            $this->populatePosition();
         }
         $this->info("Finish collect data...");
     }
@@ -65,6 +72,17 @@ class UpdateDataCommand extends Command
         MixTelematics::groups()->saveInDatabase(command: $this);
 
         $this->info("Finish populate 'Group' table");
+    }
+
+    private function populatePosition()
+    {
+        $this->alert("Start populate 'Position' table");
+
+        $since = CarbonImmutable::now()->startOfDay()->format('YmdHis') . '000';
+
+        MixTelematics::positions()->getPositions(sinceToken: $since,command: $this);
+
+        $this->info("Finish populate 'Position' table");
     }
 
 }