+++ src/Facades/MixTelematics.php
@@ -6,6 +6,7 @@ use Aislandener\MixTelematicsLaravel\Services\ActiveEventsService;
 use Aislandener\MixTelematicsLaravel\Services\AssetService;
 use Aislandener\MixTelematicsLaravel\Services\DriversService;
 use Aislandener\MixTelematicsLaravel\Services\GroupService;
+use Aislandener\MixTelematicsLaravel\Services\PositionService;
 use Illuminate\Support\Facades\Facade;
 
 /**
@@ -13,6 +14,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static ActiveEventsService activeEvents()
  * @method static AssetService assets()
  * @method static GroupService groups()
+ * @method static PositionService positions()
  */
 class MixTelematics extends Facade
 {