+++ src/MixTelematicsLaravelService.php
@@ -7,6 +7,7 @@ use Aislandener\MixTelematicsLaravel\Services\ActiveEventsService;
 use Aislandener\MixTelematicsLaravel\Services\AssetService;
 use Aislandener\MixTelematicsLaravel\Services\DriversService;
 use Aislandener\MixTelematicsLaravel\Services\GroupService;
+use Aislandener\MixTelematicsLaravel\Services\PositionService;
 use Carbon\Carbon;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Http;
@@ -20,16 +21,20 @@ class MixTelematicsLaravelService
     private DriversService $driversService;
     private AssetService $assetService;
     private GroupService $groupService;
+    private PositionService $positionService;
 
     public function __construct(ActiveEventsService $activeEventsService,
-                                DriversService $driversService,
-                                AssetService $assetService,
-                                GroupService $groupService)
+                                DriversService      $driversService,
+                                AssetService        $assetService,
+                                GroupService        $groupService,
+                                PositionService     $positionService,
+    )
     {
         $this->driversService = $driversService;
         $this->activeEventsService = $activeEventsService;
         $this->assetService = $assetService;
         $this->groupService = $groupService;
+        $this->positionService = $positionService;
     }
 
     /**
@@ -62,6 +67,11 @@ class MixTelematicsLaravelService
     public function groups(): GroupService
     {
         return $this->groupService;
+    }
+
+    public function positions(): PositionService
+    {
+        return $this->positionService;
     }