+++ src/Models/Asset.php
@@ -5,6 +5,7 @@ namespace Aislandener\MixTelematicsLaravel\Models;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasOne;
 
 class Asset extends Model
 {
@@ -63,6 +64,11 @@ class Asset extends Model
     public function group(): BelongsTo
     {
         return $this->belongsTo(Group::class,'SiteId','GroupId');
+    }
+
+    public function position() : HasOne
+    {
+        return $this->hasOne(Position::class, 'AssetId', 'AssetId');
     }
 
 }